Build a commonoption for all managers
