+++ pkg/cmd/hub/addon.go
@@ -1,20 +1,26 @@
 package hub
 
 import (
-	"github.com/openshift/library-go/pkg/controller/controllercmd"
+	"context"
+
 	"github.com/spf13/cobra"
 
 	"open-cluster-management.io/ocm/pkg/addon"
+	commonoptions "open-cluster-management.io/ocm/pkg/common/options"
 	"open-cluster-management.io/ocm/pkg/version"
 )
 
 // NewAddonManager generates a command to start addon manager
 func NewAddonManager() *cobra.Command {
-	cmdConfig := controllercmd.
+	opts := commonoptions.NewOptions()
+	cmdConfig := opts.
 		NewControllerCommandConfig("manager", version.Get(), addon.RunManager)
-	cmd := cmdConfig.NewCommand()
+	cmd := cmdConfig.NewCommandWithContext(context.TODO())
 	cmd.Use = "manager"
 	cmd.Short = "Start the Addon Manager"
+
+	flags := cmd.Flags()
+	opts.AddFlags(flags)
 
 	return cmd
 }
