+++ pkg/cmd/hub/placement.go
@@ -1,19 +1,25 @@
 package hub
 
 import (
-	"github.com/openshift/library-go/pkg/controller/controllercmd"
+	"context"
+
 	"github.com/spf13/cobra"
 
+	commonoptions "open-cluster-management.io/ocm/pkg/common/options"
 	controllers "open-cluster-management.io/ocm/pkg/placement/controllers"
 	"open-cluster-management.io/ocm/pkg/version"
 )
 
 func NewPlacementController() *cobra.Command {
-	cmd := controllercmd.
-		NewControllerCommandConfig("placement", version.Get(), controllers.RunControllerManager).
-		NewCommand()
+	opts := commonoptions.NewOptions()
+	cmdConfig := opts.
+		NewControllerCommandConfig("placement", version.Get(), controllers.RunControllerManager)
+	cmd := cmdConfig.NewCommandWithContext(context.TODO())
 	cmd.Use = "controller"
 	cmd.Short = "Start the Placement Scheduling Controller"
+
+	flags := cmd.Flags()
+	opts.AddFlags(flags)
 
 	return cmd
 }
