+++ pkg/cmd/spoke/registration.go
@@ -3,7 +3,6 @@ package spoke
 import (
 	"context"
 
-	"github.com/openshift/library-go/pkg/controller/controllercmd"
 	"github.com/spf13/cobra"
 	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
 
@@ -19,7 +18,7 @@ func NewRegistrationAgent() *cobra.Command {
 	agentOptions := spoke.NewSpokeAgentOptions()
 	commonOptions := commonoptions.NewAgentOptions()
 	cfg := spoke.NewSpokeAgentConfig(commonOptions, agentOptions)
-	cmdConfig := controllercmd.
+	cmdConfig := commonOptions.CommoOpts.
 		NewControllerCommandConfig("registration-agent", version.Get(), cfg.RunSpokeAgent)
 
 	cmd := cmdConfig.NewCommandWithContext(context.TODO())
@@ -32,7 +31,5 @@ func NewRegistrationAgent() *cobra.Command {
 
 	utilruntime.Must(features.SpokeMutableFeatureGate.Add(ocmfeature.DefaultSpokeRegistrationFeatureGates))
 	features.SpokeMutableFeatureGate.AddFlag(flags)
-
-	flags.BoolVar(&cmdConfig.DisableLeaderElection, "disable-leader-election", false, "Disable leader election for the agent.")
 	return cmd
 }
